Split import composition validation
